deployment/requirements.txt
streamlit
pandas
seaborn
matplotlib
plotly
Pillow
scikit-learn==1.2.2
tensorflow
opencv-python-headless